Move cacheResults function to helpers
