+++ app/frontend/helpers/cacheResults.ts
@@ -1,0 +1,18 @@
+import getHash from './getHash'
+
+const cacheResults = (maxAge: number, cache_obj: Object = {}) => <T extends Function>(fn: T): T => {
+  const wrapped = (...args) => {
+    const hash = getHash(JSON.stringify(args))
+    if (!cache_obj[hash] || cache_obj[hash].timestamp + maxAge < Date.now()) {
+      cache_obj[hash] = {
+        timestamp: Date.now(),
+        data: fn(...args),
+      }
+    }
+    return cache_obj[hash].data
+  }
+
+  return (wrapped as any) as T
+}
+
+export default cacheResults
